requirements.txt
requests
mysql-connector-python>=8.0.0